# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 강아지 팬카페 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Post {
  id: number;
  authorId: string;
  authorName: string;
  title: string;
  content: string;
  /** @nullable */
  imageUrl: string | null;
  likeCount: number;
  commentCount: number;
  createdAt: string;
}

export interface PostList {
  posts: Post[];
  total: number;
  page: number;
  limit: number;
}

export interface CreatePostBody {
  title: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Comment {
  id: number;
  postId: number;
  authorId: string;
  authorName: string;
  content: string;
  createdAt: string;
}

export interface CreateCommentBody {
  content: string;
}

export interface LikeResult {
  liked: boolean;
  likeCount: number;
}

export type TimelineEventMediaType =
  (typeof TimelineEventMediaType)[keyof typeof TimelineEventMediaType];

export const TimelineEventMediaType = {
  image: "image",
  video: "video",
  none: "none",
} as const;

export interface TimelineEvent {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  mediaUrl: string | null;
  mediaType: TimelineEventMediaType;
  eventDate: string;
  createdAt: string;
}

export type CreateTimelineEventBodyMediaType =
  (typeof CreateTimelineEventBodyMediaType)[keyof typeof CreateTimelineEventBodyMediaType];

export const CreateTimelineEventBodyMediaType = {
  image: "image",
  video: "video",
  none: "none",
} as const;

export interface CreateTimelineEventBody {
  title: string;
  description: string;
  /** @nullable */
  mediaUrl?: string | null;
  mediaType: CreateTimelineEventBodyMediaType;
  eventDate: string;
}

export type ProfileSocialLinks = {
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  youtube?: string | null;
  /** @nullable */
  blog?: string | null;
};

export interface Profile {
  dogName: string;
  dogBreed: string;
  dogBirthdate: string;
  dogDescription: string;
  dogImageUrl: string;
  ownerName: string;
  ownerBio: string;
  ownerImageUrl: string;
  socialLinks: ProfileSocialLinks;
}

export interface Stats {
  totalPosts: number;
  totalMembers: number;
  totalLikes: number;
  totalComments: number;
}

export type ListPostsParams = {
  page?: number;
  limit?: number;
};