# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 강아지 팬카페 API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: posts
    description: Community posts
  - name: comments
    description: Post comments
  - name: timeline
    description: Dog life timeline
  - name: profile
    description: Owner profile
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List community posts
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostList"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Unauthorized

  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Toggle like on a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
        "401":
          description: Unauthorized

  /posts/{id}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments on a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Create a comment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized

  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /timeline:
    get:
      operationId: listTimeline
      tags: [timeline]
      summary: Get dog life timeline events
      responses:
        "200":
          description: Timeline events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"
    post:
      operationId: createTimelineEvent
      tags: [timeline]
      summary: Create a timeline event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimelineEventBody"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineEvent"
        "401":
          description: Unauthorized

  /timeline/{id}:
    delete:
      operationId: deleteTimelineEvent
      tags: [timeline]
      summary: Delete a timeline event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get owner and dog profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  /stats:
    get:
      operationId: getStats
      tags: [posts]
      summary: Get community stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Post:
      type: object
      properties:
        id:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        likeCount:
          type: integer
        commentCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - authorId
        - authorName
        - title
        - content
        - imageUrl
        - likeCount
        - commentCount
        - createdAt

    PostList:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - posts
        - total
        - page
        - limit

    CreatePostBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
      required:
        - title
        - content

    Comment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - postId
        - authorId
        - authorName
        - content
        - createdAt

    CreateCommentBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    LikeResult:
      type: object
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer
      required:
        - liked
        - likeCount

    TimelineEvent:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        mediaUrl:
          type: ["string", "null"]
        mediaType:
          type: string
          enum: [image, video, none]
        eventDate:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - mediaUrl
        - mediaType
        - eventDate
        - createdAt

    CreateTimelineEventBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        mediaUrl:
          type: ["string", "null"]
        mediaType:
          type: string
          enum: [image, video, none]
        eventDate:
          type: string
      required:
        - title
        - description
        - mediaType
        - eventDate

    Profile:
      type: object
      properties:
        dogName:
          type: string
        dogBreed:
          type: string
        dogBirthdate:
          type: string
        dogDescription:
          type: string
        dogImageUrl:
          type: string
        ownerName:
          type: string
        ownerBio:
          type: string
        ownerImageUrl:
          type: string
        socialLinks:
          type: object
          properties:
            instagram:
              type: ["string", "null"]
            youtube:
              type: ["string", "null"]
            blog:
              type: ["string", "null"]
      required:
        - dogName
        - dogBreed
        - dogBirthdate
        - dogDescription
        - dogImageUrl
        - ownerName
        - ownerBio
        - ownerImageUrl
        - socialLinks

    Stats:
      type: object
      properties:
        totalPosts:
          type: integer
        totalMembers:
          type: integer
        totalLikes:
          type: integer
        totalComments:
          type: integer
      required:
        - totalPosts
        - totalMembers
        - totalLikes
        - totalComments