# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 강아지 팬카페 API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List community posts
 */
export const listPostsQueryPageDefault = 1;
export const listPostsQueryLimitDefault = 10;

export const ListPostsQueryParams = zod.object({
  page: zod.coerce.number().default(listPostsQueryPageDefault),
  limit: zod.coerce.number().default(listPostsQueryLimitDefault),
});

export const ListPostsResponse = zod.object({
  posts: zod.array(
    zod.object({
      id: zod.number(),
      authorId: zod.string(),
      authorName: zod.string(),
      title: zod.string(),
      content: zod.string(),
      imageUrl: zod.string().nullable(),
      likeCount: zod.number(),
      commentCount: zod.number(),
      createdAt: zod.string(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
});

/**
 * @summary Create a new post
 */
export const CreatePostBody = zod.object({
  title: zod.string(),
  content: zod.string(),
  imageUrl: zod.string().nullish(),
});

/**
 * @summary Get a post
 */
export const GetPostParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPostResponse = zod.object({
  id: zod.number(),
  authorId: zod.string(),
  authorName: zod.string(),
  title: zod.string(),
  content: zod.string(),
  imageUrl: zod.string().nullable(),
  likeCount: zod.number(),
  commentCount: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a post
 */
export const DeletePostParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Toggle like on a post
 */
export const LikePostParams = zod.object({
  id: zod.coerce.number(),
});

export const LikePostResponse = zod.object({
  liked: zod.boolean(),
  likeCount: zod.number(),
});

/**
 * @summary List comments on a post
 */
export const ListCommentsParams = zod.object({
  id: zod.coerce.number(),
});

export const ListCommentsResponseItem = zod.object({
  id: zod.number(),
  postId: zod.number(),
  authorId: zod.string(),
  authorName: zod.string(),
  content: zod.string(),
  createdAt: zod.string(),
});
export const ListCommentsResponse = zod.array(ListCommentsResponseItem);

/**
 * @summary Create a comment
 */
export const CreateCommentParams = zod.object({
  id: zod.coerce.number(),
});

export const CreateCommentBody = zod.object({
  content: zod.string(),
});

/**
 * @summary Delete a comment
 */
export const DeleteCommentParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get dog life timeline events
 */
export const ListTimelineResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  mediaUrl: zod.string().nullable(),
  mediaType: zod.enum(["image", "video", "none"]),
  eventDate: zod.string(),
  createdAt: zod.string(),
});
export const ListTimelineResponse = zod.array(ListTimelineResponseItem);

/**
 * @summary Create a timeline event
 */
export const CreateTimelineEventBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  mediaUrl: zod.string().nullish(),
  mediaType: zod.enum(["image", "video", "none"]),
  eventDate: zod.string(),
});

/**
 * @summary Delete a timeline event
 */
export const DeleteTimelineEventParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get owner and dog profile
 */
export const GetProfileResponse = zod.object({
  dogName: zod.string(),
  dogBreed: zod.string(),
  dogBirthdate: zod.string(),
  dogDescription: zod.string(),
  dogImageUrl: zod.string(),
  ownerName: zod.string(),
  ownerBio: zod.string(),
  ownerImageUrl: zod.string(),
  socialLinks: zod.object({
    instagram: zod.string().nullish(),
    youtube: zod.string().nullish(),
    blog: zod.string().nullish(),
  }),
});

/**
 * @summary Get community stats
 */
export const GetStatsResponse = zod.object({
  totalPosts: zod.number(),
  totalMembers: zod.number(),
  totalLikes: zod.number(),
  totalComments: zod.number(),
});