# lib/api-zod/src/generated/types/createTimelineEventBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 강아지 팬카페 API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTimelineEventBodyMediaType } from "./createTimelineEventBodyMediaType";

export interface CreateTimelineEventBody {
  title: string;
  description: string;
  /** @nullable */
  mediaUrl?: string | null;
  mediaType: CreateTimelineEventBodyMediaType;
  eventDate: string;
}