# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 강아지 팬카페 API
 * OpenAPI spec version: 0.1.0
 */

export interface Post {
  id: number;
  authorId: string;
  authorName: string;
  title: string;
  content: string;
  /** @nullable */
  imageUrl: string | null;
  likeCount: number;
  commentCount: number;
  createdAt: string;
}