# lib/api-zod/src/generated/types/postList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 강아지 팬카페 API
 * OpenAPI spec version: 0.1.0
 */
import type { Post } from "./post";

export interface PostList {
  posts: Post[];
  total: number;
  page: number;
  limit: number;
}