# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 강아지 팬카페 API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileSocialLinks } from "./profileSocialLinks";

export interface Profile {
  dogName: string;
  dogBreed: string;
  dogBirthdate: string;
  dogDescription: string;
  dogImageUrl: string;
  ownerName: string;
  ownerBio: string;
  ownerImageUrl: string;
  socialLinks: ProfileSocialLinks;
}