# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 강아지 팬카페 API
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  totalPosts: number;
  totalMembers: number;
  totalLikes: number;
  totalComments: number;
}